WEB3-103: chore: use block hash for the provider (#218)
The `AlloyDb` is created for a specific block, and represents / queries
the state with respect to that block.
This PR changes that reference from a block number to a block hash,
which should prevent problems in weird situations like re-orgs or
misconfigured RPCs.
This PR only changes the underlying representation, not the Steel API.