WEB3-94: Updated deployment scripts and guide from 1.1 deployment (#211)
During the 1.1 deployment, I implemented a number of improvements to the
deployment process to make it a bit easier and less error prone.
Primary among these changes is adding a `deployment.toml` file that
records the addresses of and associated information about each chain we
deploy to, and the contracts that are deployed there.
My hope is to use this file to do additional work around further
automating deployment, testing that deployments are correctly configured
on-chain, and generating the table of addresses that goes into our docs.