WEB3-112: chore: Steel e2e tests (#228)
This PR turns the `erc20-counter` example into an end-to-end test. It
uses a local Ethereum devnet including an execution and consensus client
based on the
[ethpandaops/ethereum-package](https://github.com/ethpandaops/ethereum-package)
Kurtosis package.
In the CI, all contracts are deployed on the devnet before a full proof
is generated and validated with a block and a beacon commitment.

closes #203
closes WEB3-90

---------

Co-authored-by: Victor Graf <victor@risczero.com>