WEB3-144: Fix incorrect encoding used by prove cheatcode (#254)
With the change from `ethers` to `alloy`, we made an error in assuming
they'd handle ABI encoding of `Vec<_>` the same. This broke the `prove`
cheatcode that is used in the Foundry template tests. We did not catch
this error prior to releasing, because we had moved all our tests in
this repo over to using `RiscZeroMockVerifier.mockProve` instead and
forgot to add test coverage specifically for this cheatcode.

This PR fixes the encoding issue and additionally adds a simple test for
using `risc0-forge-ffi` in dev mode.

Additionally, this PR solves a problem where the FFI cheatcode would
fail if `RUST_LOG` is set.

---------

Co-authored-by: Wolfgang Welz <welzwo@gmail.com>