WEB3-152: Generalized beacon commits (#265)
This PR changes two aspects of how beacon block root commits are
implemented:
- Instead of having a `const` specifying the generalized Merkle index,
we introduce a const generic for the index to make it part of the
struct. This makes the code more robust against future protocol
upgrades, and would also allow us to verify fields other than the
`block_hash` in the future.
- Instead of committing to the root of a beacon block referenced by its
timestamp, we commit to the root of a beacon block referenced by its
child timestamp, or equivalently, we commit to the root of the parent
beacon block referenced by its timestamp. While this may sound
counterintuitive, this is exactly how the
[EIP-4788](https://eips.ethereum.org/EIPS/eip-4788) beacon root contract
stores its data. This makes the verification side in Solidity much
easier and less expensive, and gets rid of the weird code that was
necessary to query the child of a beacon block during creation.

This PR also adds much more documentation explaining that beacon
commitment creation is currently only implemented for Daneb, and that
the library needs to be updated to support future versions.

It also introduces a wrapper around the Eth RPC block number tags,
making it more convenient to specify the last block in the chain that
still has a child.

Resolves #258, resolves #260, resolves #262
Resolves WEB3-150, resolves WEB3-152, resolves WEB3-154
This PR also resolves #242 and resolves WEB3-123, as without the need to
query for the child beacon node that api is not needed.