WEB3-213: add `SteelVerifier` to verify Steel commitments in the guest (#319)
This PR introduces the `SteelVerifier`, which acts as a built-in Steel
`Contract` to verify Steel commitments. It is used like any other
`Contract`, during the preflight step and in the guest. This
functionality is currently marked unstable and must be enabled using the
`unstable-verifier` feature.

The `TokenStats' example has been updated to use the verifier to sample
the stats over two datapoints at different times.

---------

Co-authored-by: Victor Snyder-Graf <victor@risczero.com>