BM-85: Refactor aggregation set builder for incremental tree building (#381)
The aggregation set builder is used extensively in Boundless and one of
the main bottlenecks right now with service large numbers of small
requests is that aggregation is done in a series of sequential steps.
In theory, these steps could be parallelized but this is complicated in
practice.

This PR refactors the set-builder to be able to accept multiple claims
per execution, and can aggregated a given vector of receipts in one-step
if immediate finalization is desired.
It does so by using a Merkle mountain range structure to encode
intermediate states in building the final tree.
When combined with the work on parallelized assumption resolution
(through the `union` program) this should take advantage of existing
parallelized workflows in the prover cluster to achieve much better e2e
times on receipt aggregation, and therefore on request fulfillment.

Related: BM-85

---------

Co-authored-by: Angelo Capossele <angelocapossele@gmail.com>
Co-authored-by: Wolfgang Welz <welzwo@gmail.com>