suggestions for the steel-history.md doc (#388)
Note on how I broke up paragraphs onto multiple lines. It's not a huge
benefit in this PR, but splitting up paragraphs helps with reviews for
changes later because the diff will show the changes in a nicer way. In
prose like this, I prefer to break up paragraphs with one sentence or
clause per line. I find that works well for collaborative editing.