WEB3-398: Use tagged leaf hashing in aggregation set to improve missue resistance (#526)
On `main` the `RiscZeroSetVerifier.verifyIntegrity` function will accept
internal nodes in the Merkle tree, and complete with reverting. This is
not an intended use of this interface, and is a violation of the calling
requirements as the passed hash is not the digest of a claim (this is
similar to running ECDSA's verify operation on an unhashed message).
However, this is misuse that can be mitigated by following the best
practice of tagging Merkle tree leaves.

This PR makes the aggregation set verifier more misuse resistant by
tagging and hashing the leaves of the tree.

Additionally, the documentation on `IRiscZeroVerifier.verifyIntegrity`
is updated to highlight the importance of the caller hashing the claim
digest and to be more precise about the guarantees provided.

Resolves: https://github.com/risc0/risc0-ethereum/issues/525
Resolves: WEB3-397