WEB3-486: Update deployment.toml to reflect stopped 2.0 contracts (#628)
Following the stopping of the 2.0 verifiers on the official router [1],
this PR updated the deployment.toml file to reflect this, and improve
the deployment tests to test the state the estop using a receipt
verification. To support this, a 2.0 receipt is added along with
functionality to test verification of multiple versions of the zkVM
verifier.

[1]
https://github.com/risc0/risc0/security/advisories/GHSA-g3qg-6746-3mg9