WEB3-525: Create a simplified contract integration test, that does not attempt to test Steel (#667)
Our existing e2e test involves running a full devnet for an Ethereum
beacon chain. This is done in order to have the e2e test include Steel.
With Steel moving to its own repo, we no longer need to do / can do
this.

This PR includes a new `test-utils` crate, and includes in it a simpler
e2e test that includes the prover, including Bonsai, and the contracts.